feat(ui): milestones UI is simplefied
