+++ src/app/body/create-new-task/create-new-task.component.html
@@ -157,9 +157,8 @@
                     </ng-container>
 
                 
-                <div class="modal-footer mt-n1">
+                <div class="modal-footer mt-1">
                     <button type="submit" class="btn btn-lg btn-block" id="button" [disabled]="enableLoader" aria-label="Close" (click)="submit()">Create New Task</button>
-                    
                 </div>
                 <app-loader *ngIf="enableLoader"></app-loader>
                     </form>
