+++ src/app/body/quick-notes/quick-notes.component.css
@@ -37,7 +37,7 @@
 }
 
 .btn {
-    color: var(--text);
+    color: #ffffff;
 }
 
 .btn:hover {
