+++ src/app/body/quick-notes/quick-notes.component.html
@@ -12,7 +12,7 @@
 * See the MIT License for more details. 
 ***********************************************************/ -->
 <button class="btn btn-success" id="quickNotesButton" (click)="showList()">
-    <mat-icon fontSet="material-icons" class="align-middle">
+    <mat-icon fontSet="material-icons-outlined" class="align-middle">
         notes
     </mat-icon>
     Quick Note
