+++ src/app/icons/generate-icon/generate-icon.component.css
@@ -1,10 +1,9 @@
 .icon {
     padding: right 16px;
-    
 }
+
 .Border {
     text-align: center;
     border-style: dashed;
-    border-width: 1px;  
-    padding-top: 4px;    
+    border-width: 1px;
 }
